refatoring: move - Function - generateNotification(ResultSet result, Notification notification1) origin (NotificationDAO) -> destiny (NotificationFactory)
